admin time & clock last4 screen done

+++ android/app/build.gradle.kts
@@ -27,7 +27,8 @@ android {
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
 
-        minSdkVersion flutter.minSdkVersion
+        
+        minSdkVersion(flutter.minSdkVersion.toInt())
 
         targetSdk = 36
         versionCode = flutter.versionCode
